+++ configs/default_logging_config.conf
@@ -1,0 +1,33 @@
+[loggers]
+keys=root
+
+[handlers]
+keys=consoleHandler,fileHandler
+
+[formatters]
+keys=consoleFormatter,fileFormatter
+
+[logger_root]
+handlers = consoleHandler, fileHandler
+level = INFO
+qualname = rootLogger
+propagate = 0
+
+[handler_consoleHandler]
+class = logging.StreamHandler
+level = INFO
+formatter = consoleFormatter
+
+[formatter_consoleFormatter]
+format = %(asctime)s : %(message)s
+datefmt = %d-%m-%Y %I:%M:%S
+
+[handler_fileHandler]
+class=logging.FileHandler
+level=DEBUG
+formatter=fileFormatter
+args=("%(logfilename)s", "w")
+
+[formatter_fileFormatter]
+format = %(asctime)s : %(levelname)s : %(module)s : %(funcName)s : %(lineno)d : (Process Details : (%(process)d, %(processName)s), Thread Details : (%(thread)d, %(threadName)s))\nLog : %(message)s
+datefmt = %d-%m-%Y %I:%M:%S
