+++ src/wai_data_tools/scripts/manually_reclassify_frames.py
@@ -6,7 +6,7 @@ import pathlib
 import pandas as pd
 import tqdm
 
-import wai_data_tools.io
+import wai_data_tools.file_handling
 from wai_data_tools import config, manual_labeling
 
 
@@ -34,7 +34,7 @@ def manually_reclassify_frames(
     logging.info("Starting GUI for reclassification")
     video_dirs = [dir_path for dir_path in dataset_dir.iterdir() if dir_path.is_dir()]
     for video_dir in tqdm.tqdm(video_dirs):
-        frames_dict = wai_data_tools.io.load_frames(frame_dir=video_dir, df_frames=df_frames)
+        frames_dict = wai_data_tools.file_handling.load_frames(frame_dir=video_dir, df_frames=df_frames)
         manual_labeling.manual_annotation_plot(
             frame_dict=frames_dict,
             df_frames=df_frames,
